06.Agents/U04/pyproject.toml
[project]
name = "u04"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mcp[cli]>=1.26.0",
    "ollama>=0.6.1",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.6",
]
